fixing issues

+++ RedifyQuiz/Utils/Utilities.cs
@@ -10,7 +10,7 @@ namespace RedifyQuiz.Utils
         /// </summary>
         /// <param name="n"></param>
         /// <returns>integer</returns>
-        public static int NthFibonacciNumber(int n)
+        public static int NthFibonacciNumber(double n)
         {
             int a = 0;
             int b = 1;
